Python/MySite/Tools/requirements.txt
absl-py==1.0.0
algorithms==0.1.4
altgraph==0.17.2
alyn==0.1.1
argon2-cffi==21.1.0
arrow==1.1.1
asgiref==3.4.1
astunparse==1.6.3
attrs==21.2.0
auto-py-to-exe==2.11.0
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
binaryornot==0.4.4
bleach==4.1.0
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.9
click==7.1.2
colorama==0.4.4
comtypes==1.1.10
cookiecutter==1.7.3
cookiejar==0.0.3
cryptography==3.4.8
cycler==0.10.0
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
deskew==0.10.36
distlib==0.3.3
Django==3.2.7
docopt==0.6.2
docx==0.2.4
Eel==0.12.4
eng-to-ipa==0.0.2
entrypoints==0.3
et-xmlfile==1.1.0
evernote==1.19
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.3.2
find==2020.12.3
flashtext==2.7
flatbuffers==2.0
future==0.18.2
gast==0.4.0
gcloud==0.18.3
gevent==21.8.0
gevent-websocket==0.10.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
googletrans==3.0.0
greenlet==1.1.2
grpcio==1.42.0
gTTS==2.2.3
h11==0.9.0
h2==3.2.0
h5py==3.6.0
hpack==3.0.0
hstspreload==2021.9.1
html5lib==1.1
httpcore==0.9.1
httplib2==0.19.1
httpx==0.13.3
hyperframe==5.2.0
idna==3.3
imageio==2.9.0
importlib-metadata==4.8.2
ipykernel==4.6.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
iso8601==0.1.16
jedi==0.18.0
Jinja2==3.0.1
jinja2-time==0.2.0
joblib==1.1.0
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
jwcrypto==1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
keyboard==0.13.5
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.6.4
m3u8==0.9.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mccabe==0.6.1
mechanize==0.4.6
mistune==0.8.4
MouseInfo==0.1.3
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.5
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python==4.5.4.58
openpyxl==3.0.8
opt-einsum==3.3.0
packaging==21.0
pager==3.3
pandas==1.3.2
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pdf2image==1.16.0
pdfminer.six==20200517
PDFNetPython3==9.1.0
pdfplumber==0.5.28
pefile==2021.9.3
pickleshare==0.7.5
Pillow==8.4.0
pipreqs==0.4.10
platformdirs==2.4.0
plotly==5.3.1
portpicker==1.2.0
poyo==0.5.0
print-dict==0.1.19
prometheus-client==0.12.0
prompt-toolkit==3.0.21
protobuf==3.19.1
py2exe==0.10.4.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.10.1
pyflakes==2.4.0
pygame==2.0.2
pygame-2048==1.9.4
PyGetWindow==0.0.9
Pygments==2.10.0
pyinstaller==4.6
pyinstaller-hooks-contrib==2021.3
PyMsgBox==1.0.9
PyMuPDF==1.19.1
pynput==1.7.5
pyodbc==4.0.32
pyparsing==2.4.7
PyPDF2==1.26.0
pypdfocr==0.9.1
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2
PyQt5Designer==5.14.1
Pyrebase4==4.5.0
PyRect==0.1.4
pyrsistent==0.18.0
PyScreeze==0.1.28
PySocks==1.7.1
pytesseract==0.3.8
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.2
python-jwt==3.3.0
python-slugify==5.0.2
pyttsx3==2.90
pytweening==1.0.4
pytz==2021.1
PyWavelets==1.1.1
pywin32==301
pywin32-ctypes==0.2.0
pywinpty==1.1.5
pyworld3==1.1
pyzmq==22.3.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.10.23
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.8
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
simplegeneric==0.8.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
terminado==0.12.1
tesseract==0.1.3
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.0.0
tifffile==2021.10.12
tkPDFViewer==0.1
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.7
virtualenv==20.9.0
Wand==0.6.7
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
whichcraft==0.6.1
widgetsnbextension==3.5.2
wrapt==1.13.3
wxPython==4.1.1
XlsxWriter==3.0.2
xlwings==0.24.9
yapf==0.30.0
yarg==0.1.9
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
